+++ Base/LogDefault.php
@@ -1,0 +1,24 @@
+<?php
+namespace TijsVerkoyen\DocDataPayments\Base;
+class LogDefault implements LogInterface
+{
+    public function info($message, $data = null)
+    {
+        // TODO: Implement info() method.
+    }
+
+    public function warning($message, $data = null)
+    {
+        // TODO: Implement warning() method.
+    }
+
+    public function error($message, $data = null)
+    {
+        // TODO: Implement error() method.
+    }
+
+    public function getLog()
+    {
+        // TODO: Implement error() method.
+    }
+}