+++ DocDataPayments.php
@@ -1,6 +1,8 @@
 <?php
 namespace TijsVerkoyen\DocDataPayments;
 
+use TijsVerkoyen\DocDataPayments\Base\LogDefault;
+use TijsVerkoyen\DocDataPayments\Base\LogInterface;
 use TijsVerkoyen\DocDataPayments\Types\Amount;
 use TijsVerkoyen\DocDataPayments\Types\CancelRequest;
 use TijsVerkoyen\DocDataPayments\Types\CaptureRequest;
@@ -16,6 +18,7 @@ use TijsVerkoyen\DocDataPayments\Types\RefundRequest;
 use TijsVerkoyen\DocDataPayments\Types\SepaBankAccount;
 use TijsVerkoyen\DocDataPayments\Types\Shopper;
 use TijsVerkoyen\DocDataPayments\Types\StatusRequest;
+use TijsVerkoyen\DocDataPayments\Types\PaidLevel;
 
 /**
  * Docdata Payments class
@@ -36,11 +39,18 @@ class DocDataPayments
     private $merchant;
 
     /**
-     * The soapclient
+ * The soapclient
+ *
+ * @var SoapClient
+ */
+    private $soapClient;
+
+    /**
+     * Logger
      *
-     * @var \SoapClient
-     */
-    private $soapclient;
+     * @var LogInterface
+    */
+    public $logger;
 
     /**
      * The timeout
@@ -120,11 +130,13 @@ class DocDataPayments
      *
      * @param string[optional] $wsdl The location of the WSDL-file
      */
-    public function __construct($wsdl)
+    public function __construct($wsdl, LogInterface $logger = null)
     {
         if ($wsdl !== null) {
             $this->setWsdl($wsdl);
         }
+
+        $this->logger = ($logger == null) ? new LogDefault() : $logger;
     }
 
     /**
@@ -135,9 +147,9 @@ class DocDataPayments
     protected function getSoapClient()
     {
         // create the client if needed
-        if (!$this->soapclient) {
+        if (!$this->soapClient) {
             $options = array(
-                'trace' => self::DEBUG,
+                'trace' => true,
                 'exceptions' => true,
                 'connection_timeout' => $this->getTimeout(),
                 'user_agent' => $this->getUserAgent(),
@@ -295,7 +307,10 @@ class DocDataPayments
         }
 
         // make the call
+        $this->logger->info("Payment create: " .$merchantOrderReference, $request->toArray());
         $response = $this->getSoapClient()->create($request->toArray());
+        $this->logger->info("Payment create soap request: " .$merchantOrderReference, $this->soapClient->__getLastRequest());
+        $this->logger->info("Payment create soap response: " .$merchantOrderReference, $this->soapClient->__getLastResponse());
 
         // validate response
         if (isset($response->createError)) {
@@ -303,6 +318,9 @@ class DocDataPayments
                 var_dump($this->soapClient->__getLastRequest());
                 var_dump($response->createError);
             }
+
+            $this->logger->error("Payment create: " .$merchantOrderReference, $response->createError->getError()->getExplanation());
+
 
             throw new Exception(
                 $response->createError->getError()->getExplanation()
@@ -327,7 +345,13 @@ class DocDataPayments
         $request->setPaymentOrderKey($paymentOrderKey);
 
         // make the call
+
+
+        $this->logger->info("Payment cancel: " .$paymentOrderKey, $request->toArray());
         $response = $this->getSoapClient()->cancel($request->toArray());
+        $this->logger->info("Payment cancel soap request: " .$paymentOrderKey, $this->soapClient->__getLastRequest());
+        $this->logger->info("Payment cancel soap response: " .$paymentOrderKey, $this->soapClient->__getLastResponse());
+
 
         // validate response
         if (isset($response->cancelError)) {
@@ -335,6 +359,7 @@ class DocDataPayments
                 var_dump($this->soapClient->__getLastRequest());
                 var_dump($response->cancelError);
             }
+            $this->logger->error("Payment cancel: " .$paymentOrderKey, $response->cancelError->getError()->getExplanation());
 
             throw new Exception(
                 $response->cancelError->getError()->getExplanation()
@@ -400,7 +425,11 @@ class DocDataPayments
         }
 
         // make the call
+
+        $this->logger->info("Payment capture: " .$merchantCaptureReference, $request->toArray());
         $response = $this->getSoapClient()->capture($request->toArray());
+        $this->logger->info("Payment capture soap request: " .$merchantCaptureReference, $this->soapClient->__getLastRequest());
+        $this->logger->info("Payment capture soap response: " .$merchantCaptureReference, $this->soapClient->__getLastResponse());
 
         // validate response
         if (isset($response->captureError)) {
@@ -408,6 +437,8 @@ class DocDataPayments
                 var_dump($this->soapClient->__getLastRequest());
                 var_dump($response->captureError);
             }
+
+            $this->logger->error("Payment capture: " .$merchantCaptureReference, $response->captureError->getError()->getExplanation());
 
             throw new Exception(
                 $response->captureError->getError()->getExplanation()
@@ -469,7 +500,11 @@ class DocDataPayments
         }
 
         // make the call
+        $this->logger->info("Payment capture: " .$merchantRefundReference, $request->toArray());
         $response = $this->getSoapClient()->refund($request->toArray());
+        $this->logger->info("Payment capture soap request: " .$merchantRefundReference, $this->soapClient->__getLastRequest());
+        $this->logger->info("Payment capture soap response: " .$merchantRefundReference, $this->soapClient->__getLastRequest());
+
 
         // validate response
         if (isset($response->refundError)) {
@@ -477,6 +512,8 @@ class DocDataPayments
                 var_dump($this->soapClient->__getLastRequest());
                 var_dump($response->refundError);
             }
+
+            $this->logger->error("Payment capture: " .$merchantRefundReference, $response->refundError->getError()->getExplanation());
 
             throw new Exception(
                 $response->refundError->getError()->getExplanation()
@@ -491,7 +528,7 @@ class DocDataPayments
      * captures or refunds. It can be used to determine whether an order is considered paid, to retrieve a payment ID,
      * to get information on the statuses of captures/refunds.
      *
-     * @param  string                                           $paymentOrderKey
+     * @param  string $paymentOrderKey
      * @return TijsVerkoyen\DocDataPayments\Types\StatusSuccess
      */
     public function status($paymentOrderKey)
@@ -501,7 +538,28 @@ class DocDataPayments
         $request->setPaymentOrderKey($paymentOrderKey);
 
         // make the call
+        $response = $this->statusReponse($paymentOrderKey);
+
+        return $response->statusSuccess;
+    }
+
+    /**
+     * The soap status response of a Payment Order,
+     *
+     * @param  string  $paymentOrderKey
+     * @return Soap Response
+     */
+    public function statusReponse($paymentOrderKey)
+    {
+        $request = new StatusRequest();
+        $request->setMerchant($this->merchant);
+        $request->setPaymentOrderKey($paymentOrderKey);
+
+        // make the call
+        $this->logger->info("Payment status: " .$paymentOrderKey, $request->toArray());
         $response = $this->getSoapClient()->status($request->toArray());
+        $this->logger->info("Payment status soap request: " .$paymentOrderKey, $this->soapClient->__getLastRequest());
+        $this->logger->info("Payment status soap response: " .$paymentOrderKey, $this->soapClient->__getLastRequest());
 
         // validate response
         if (isset($response->statusError)) {
@@ -510,12 +568,14 @@ class DocDataPayments
                 var_dump($response->statusError);
             }
 
+            $this->logger->error("Payment status: " .$paymentOrderKey, $response->statusError->getError()->getExplanation());
+
             throw new Exception(
                 $response->statusError->getError()->getExplanation()
             );
         }
 
-        return $response->statusSuccess;
+        return $response;
     }
 
     /**
@@ -575,9 +635,9 @@ class DocDataPayments
         }
 
         if ($production) {
-            $base = 'https://www.docdatapayments.com/ps/com.tripledeal.paymentservice.servlets.PaymentService';
+            $base = 'https://secure.docdatapayments.com/ps/menu';
         } else {
-            $base = 'https://test.docdatapayments.com/ps/com.tripledeal.paymentservice.servlets.PaymentService';
+            $base = 'https://test.docdatapayments.com/ps/menu';
         }
 
         // build the url
@@ -627,8 +687,62 @@ class DocDataPayments
             $production
         );
 
+        $this->logger->info("Redirect to docdata: " .$url);
+
         // redirect
         header('location: ' . $url);
         exit();
+    }
+
+    /**
+     * Docdata document: 733126_Integration_manual_Order_Api_1-1.pdf
+     * Chapter: 7.4 Determining whether an order is paid
+     * Determining whether an order is paid
+     * Different merchants can have different ways of determining when they consider an order “paid”,
+     * the totals in the status report are there to help make this decision. Keep in mind that the status report
+     * never reports about money actually having been transferred to a merchant, so it is not a complete guarantee
+     * that a payment has been finished in that sense.
+     * Using the totals to determine a level of confidence:
+     * @param $paymentOrderKey
+     * @param $currencyAmountDiffAcceptedRange int Currency difference between Euro and Dollars, this is the accepted range
+     * @return $paidlevel, Constant of the asbtract class TijsVerkoyen\DocDataPayments\Types\PaidLevel
+     */
+    public function statusPaid($paymentOrderKey, $currencyAmountDiffAcceptedRange = 0)
+    {
+        $response = $this->statusReponse($paymentOrderKey);
+
+        if(isset($response->statusSuccess) &&
+           $response->statusSuccess->getSuccess() != null &&
+           $response->statusSuccess->getSuccess()->getCode() == 'SUCCESS' &&
+           $response->statusSuccess->getReport() != null &&
+           $response->statusSuccess->getReport()->getApproximateTotals() != null
+          )
+        {
+            $approximateTotals = $response->statusSuccess->getReport()->getApproximateTotals();
+
+            //Safe Route
+            if(($approximateTotals->getTotalRegistered() <= $approximateTotals->getTotalCaptured() + $currencyAmountDiffAcceptedRange))
+            {
+                return PaidLevel::SafeRoute;
+            }
+
+            //Balanced Route
+            if($approximateTotals->getTotalRegistered()  <=
+                $approximateTotals->gettotalAcquirerApproved() + $currencyAmountDiffAcceptedRange)
+            {
+                return PaidLevel::BalancedRoute;
+            }
+
+            //Quick Route
+            if($approximateTotals->getTotalRegistered()  <=
+                    $approximateTotals->getTotalShopperPending()
+                    + $approximateTotals->getTotalAcquirerPending()
+                    + $approximateTotals->gettotalAcquirerApproved() + $currencyAmountDiffAcceptedRange)
+            {
+                return PaidLevel::QuickRoute;
+            }
+        }
+
+        return PaidLevel::NotPaid;
     }
 }