+++ Types/CaptureRequest.php
@@ -82,7 +82,7 @@ class CaptureRequest extends Object
     /**
      * @var string
      */
-    protected $version = '1.0';
+    protected $version = '1.2';
 
     /**
      * @param \TijsVerkoyen\DocDataPayments\Types\Amount $amount