+++ Types/CreateRequest.php
@@ -102,7 +102,7 @@ class CreateRequest extends Object
     /**
      * @var string
      */
-    protected $version = '1.0';
+    protected $version = '1.2';
 
     /**
      * @param \TijsVerkoyen\DocDataPayments\Types\Destination $billTo