+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Changelog since 3.0.0
+* Updated to Docdata payment Api 1.1
+  
+  Changes from Order API version 1.0
+  The extended status message has been changed. Payment methods have now been grouped in a more generic way.
+  Please refer to our XSD to get the latest definitions.
+  To activate the extended status message for your merchant account,
+  please contact Docdata Payments Customer Service.
+* New Method StatusPaid
+* 
 # Changelog since 2.0.2
 
 * Implemented the capture-method.
@@ -15,3 +25,4 @@
 # Changelog since 1.0.0
 
 * Made the class compatible with the 1.0-wsdl
+